requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blis==0.7.11
catalogue==1.0.2
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
cymem==2.0.8
dataclasses-json==0.6.7
distro==1.9.0
Django==5.1
djangorestframework==3.15.2
docx2txt==0.8
en-core-web-sm==2.3.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filelock==3.15.4
frozenlist==1.4.1
gdown==5.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.29
langchain-openai==0.1.7
langchain-postgres==0.0.9
langchain-text-splitters==0.2.2
langsmith==0.1.98
marshmallow==3.21.3
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
openai==1.40.2
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pdfminer.six==20240706
pgvector==0.2.5
plac==1.1.3
preshed==3.0.9
psycopg==3.2.1
psycopg-pool==3.2.2
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyPDF2==3.0.1
pyresparser==1.0.6
pyrsistent==0.20.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==2.3.9
SQLAlchemy==2.0.32
sqlparse==0.5.1
srsly==1.0.7
tenacity==8.5.0
thinc==7.4.6
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==0.10.1
yarl==1.9.4
